# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel",
        "description": "Joy VoyagerApi module adds REST Api end points to Voyager with Passport and Swagger support https://github.com/rxcod9/joy-voyager-api.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://finishuseruz.demo",
            "description": "Primary API Server"
        },
        {
            "url": "http://my-default-host.com",
            "description": "Another API Server"
        }
    ],
    "paths": {
...
        "/api/task/{task}": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "Show tasks by ID",
                "operationId": "f81261f5febbe5e9edecf62e8b165e8f",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/my-tasks": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "Get list of my Tasks",
                "operationId": "8d42e680a33b8ee945b3e9171b2e7c0d",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/find": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "Get list in Tasks",
                "operationId": "60b37da2619850c32aa3b6e4feccae61",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/task/create": {
            "post": {
                "tags": [
                    "Task"
                ],
                "summary": "Add new task",
                "operationId": "eaeff5e59900942953b4fdd6602d933b",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "address": {
                                                "type": "string"
                                            },
                                            "date_type": {
                                                "type": "string"
                                            },
                                            "start_date": {
                                                "type": "string"
                                            },
                                            "end_date": {
                                                "type": "string"
                                            },
                                            "budget": {
                                                "type": "integer"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "category_id": {
                                                "type": "string"
                                            },
                                            "phone": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Javoxir",
                                    "address": "Xorazm",
                                    "date_type": "1",
                                    "start_date": "2021-05-17 10:00",
                                    "end_date": "2021-05-17 10:00",
                                    "budget": 10000,
                                    "description": "Juda zo`r",
                                    "category_id": "31",
                                    "phone": "909598654"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "name"
                                        },
                                        "address": {
                                            "type": "string",
                                            "example": "address"
                                        },
                                        "date_type": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "start_date": {
                                            "type": "string",
                                            "example": "2021-05-17 00:00"
                                        },
                                        "end_date": {
                                            "type": "string",
                                            "example": "2021-05-17 00:00"
                                        },
                                        "budget": {
                                            "type": "integer",
                                            "example": "100000"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Zo`r"
                                        },
                                        "category_id": {
                                            "type": "integer",
                                            "example": "35"
                                        },
                                        "phone": {
                                            "type": "string",
                                            "example": "909598654"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "fail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/change-task/{task}": {
            "put": {
                "tags": [
                    "Task"
                ],
                "summary": "Update Task",
                "operationId": "c7d65b35593a76b8531537f6d0d373d4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "phone": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "date_type": {
                                                "type": "string"
                                            },
                                            "start_date": {
                                                "type": "string"
                                            },
                                            "budget": {
                                                "type": "integer"
                                            },
                                            "category_id": {
                                                "type": "string"
                                            },
                                            "coordinates0": {
                                                "type": "string"
                                            },
                                            "location0": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Javoxir",
                                    "phone": "+99898745623",
                                    "description": "Juda zo`r",
                                    "date_type": "1",
                                    "start_date": "2021-05-17 10:00",
                                    "budget": "10000",
                                    "category_id": "31",
                                    "coordinates0": "30.65006,35.546057",
                                    "location0": "31.65006,35.546057"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "name"
                                        },
                                        "phone": {
                                            "type": "string",
                                            "example": "+99898745623"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Zo`r"
                                        },
                                        "date_type": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "start_date": {
                                            "type": "string",
                                            "example": "2021-05-17 00:00"
                                        },
                                        "budget": {
                                            "type": "string",
                                            "example": "100000"
                                        },
                                        "category_id": {
                                            "type": "integer",
                                            "example": "35"
                                        },
                                        "coordinates0": {
                                            "type": "string",
                                            "example": "30.65006,35.546057"
                                        },
                                        "location0": {
                                            "type": "string",
                                            "example": "31.65006,35.546057"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/for_del_new_task/{task}": {
            "delete": {
                "tags": [
                    "Task"
                ],
                "summary": "Delete Task",
                "operationId": "944c123df58af87508b1e98eda756420",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "token": {
                "type": "apiKey",
                "description": "Bearer token",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}